feat(ui): implement rainbow trace visualization and training wizards
- Add GenerateDatasetWizard for automated QA/Summary generation
- Implement Rainbow Trace for per-token adapter coloring
- Add version skew banner for client-server compatibility checks
- Enhance dashboard with sparklines and improved state monitoring

+++ crates/adapteros-ui/src/components/virtual_list.rs
@@ -219,7 +219,7 @@ where
     F: Fn(T, usize) -> V + Clone + Send + Sync + 'static,
 {
     let scroll_top = RwSignal::new(0u32);
-    let container_height = (row_height * max_visible_rows as u32) as u32;
+    let container_height = row_height * max_visible_rows as u32;
 
     // Calculate the virtual range
     let virtual_range = Memo::new(move |_| {
